etl/app/requirements.txt
flake8==4.0
psycopg2
pytest==6.2.5
python-dotenv==0.20.0
requests==2.27.1
elasticsearch==8.1.0
backoff